Extract GuideThemeSelector and use it in GuideSandboxChrome.
