+++ src-docs/src/components/guide_theme_selector/index.js
@@ -1,0 +1,1 @@
+export { GuideThemeSelector } from './guide_theme_selector';
